fix remarks with constants
