+++ src/components/comment/comment.jsx
@@ -1,6 +1,20 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {Month} from '../../constants/constants';
+
+const Month = {
+  0: `January`,
+  1: `February`,
+  2: `March`,
+  3: `April`,
+  4: `May`,
+  5: `June`,
+  6: `July`,
+  7: `August`,
+  8: `September`,
+  9: `October`,
+  10: `November`,
+  11: `December`,
+};
 
 
 const Comment = (props) => {
