+++ src/components/list-comments/list-comments.jsx
@@ -2,7 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import Comment from '../comment/comment.jsx';
-import {DEFAULT_COUNT_COMMITS} from '../../constants/constants.js';
+
+const DEFAULT_COUNT_COMMITS = 10;
 
 
 const ListComments = (props) => {
@@ -22,7 +23,7 @@ const ListComments = (props) => {
       <ul className="reviews__list">
         {comments.slice(0, DEFAULT_COUNT_COMMITS).map((comment) =>
           <Comment
-            key={`comment-${comment.id}`}
+            key={comment.id}
             comment={comment}
           />
         )}
