--- src/constants/constants.js
@@ -1,21 +1,0 @@
-const Month = {
-  0: `January`,
-  1: `February`,
-  2: `March`,
-  3: `April`,
-  4: `May`,
-  5: `June`,
-  6: `July`,
-  7: `August`,
-  8: `September`,
-  9: `October`,
-  10: `November`,
-  11: `December`,
-};
-
-const DEFAULT_COUNT_COMMITS = 10;
-
-export {
-  DEFAULT_COUNT_COMMITS,
-  Month,
-};
